[DDCNL-7801] updated unit tests
